moving logger util to it's own file
